add Loaddata action
