+++ package.json
@@ -13,6 +13,7 @@
     "@babel/core": "7.9.6",
     "@babel/preset-env": "7.9.6",
     "@babel/preset-react": "7.9.4",
+    "axios-mock-adapter": "1.18.1",
     "babel-jest": "26.0.1",
     "babel-loader": "8.1.0",
     "enzyme": "3.11.0",
@@ -30,11 +31,13 @@
     "node": "10"
   },
   "dependencies": {
+    "axios": "0.19.2",
     "leaflet": "1.6.0",
     "prop-types": "15.7.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-redux": "7.2.0",
-    "redux": "4.0.5"
+    "redux": "4.0.5",
+    "redux-thunk": "2.3.0"
   }
 }
