+++ src/reducer.js
@@ -1,3 +1,37 @@
+import offers from "./mocks/data.js";
+
+const rawDataConversion = (allOffers) => {
+  let rezultArray = [];
+
+  allOffers.map((offer) => {
+    let newOffer = {
+      city: offer[`city`],
+      previewImage: offer[`preview_image`],
+      images: offer[`images`],
+      title: offer[`title`],
+      isFavorite: offer[`is_favorite`],
+      isPremium: offer[`is_premium`],
+      rating: offer[`rating`],
+      type: offer[`type`],
+      bedrooms: offer[`bedrooms`],
+      maxAdults: offer[`max_adults`],
+      price: offer[`price`],
+      goods: offer[`goods`],
+      host: {
+        id: offer[`host`][`id`],
+        name: offer[`host`][`name`],
+        isPro: offer[`host`][`is_pro`],
+        avatarUrl: offer[`host`][`avatar_url`],
+      },
+      description: offer[`description`],
+      location: offer[`location`],
+      id: offer[`id`],
+    };
+    rezultArray.push(newOffer);
+  });
+  return rezultArray;
+};
+
 const initialState = {
   city: `default`,
   data: [],
@@ -13,6 +47,11 @@ const ActionCreator = {
     type: `GET_OFFERS_LIST`,
     payload: placeOffers,
   }),
+
+  loadData: () => ({
+    type: `LOAD_DATA`,
+    payload: rawDataConversion(offers),
+  })
 };
 
 
@@ -24,6 +63,10 @@ const reducer = (state = initialState, action) => {
     });
 
     case `GET_OFFERS_LIST` : return Object.assign({}, state, {
+      data: action.payload,
+    });
+
+    case `LOAD_DATA`: return Object.assign({}, state, {
       data: action.payload,
     });
   }
